Convert price to long long

+++ src/dataManager/Order.cpp
@@ -5,12 +5,12 @@
 
 #include "exchange/Exchange.h"
 
-Order::Order(const std::string s, int p, int v, bool isBuy)
+Order::Order(const std::string s, long long p, int v, bool isBuy)
     : symbol(s), price(p), volume(v), isBuy(isBuy){};
 
 int Order::getVolume() const { return volume; }
 
-int Order::getPrice() const { return price; }
+long long Order::getPrice() const { return price; }
 
 std::string Order::getSymbol() const { return symbol; }
