+++ src/dataManager/Order.h
@@ -8,11 +8,11 @@
 
 class Order {
  public:
-  Order(std::string, int, int, bool);
-  Order(std::string symbol, int price, int volume, bool isBuyOrder,
+  Order(std::string, long long, int, bool);
+  Order(std::string symbol, long long price, int volume, bool isBuyOrder,
         Exchange::ExchangeName exchange);
   int getVolume() const;
-  int getPrice() const;
+  long long getPrice() const;
   bool isBuyOrder() const;
   std::string getSymbol() const;
   Exchange::ExchangeName getExchange();
@@ -21,7 +21,7 @@ class Order {
  private:
   Exchange::ExchangeName exchange;
   const std::string symbol;
-  const int price;
+  const long long price;
   const int volume;
   const bool isBuy;  // If not it is a sell order
 };
